Converted DateTimeManager.kt to Object and Moved UI Logic
generateListHeader() and formatZonedDate() provided UI Logic and didn't belong in Domain.  Moved the former to the VM (since it updates the state object) and the latter to AgendaScreen.kt (since it isn't applied to the state object and instead is used in the composable for each individual AgendaItem.)
