+++ app/src/main/kotlin/com/gavinferiancek/tasky/agenda/di/AgendaModule.kt
@@ -14,6 +14,6 @@ object AgendaModule {
     @Singleton
     @Provides
     fun provideDateTimeManager(): DateTimeManager {
-        return DateTimeManager()
+        return DateTimeManager
     }
 }
